+++ Source/SelfService/Web/integrations/connection/connectionDetails/consumeDataEventStreams/serviceAccounts/serviceAccountsDataGrid/viewCredentialsDialogReducer.ts
